+++ src/sparkle/CLI/help/__init__.py
@@ -1,3 +1,4 @@
 """Init for CLI help package."""
 
-from sparkle.CLI.help.nicknames import resolve_object_name, resolve_instance_name
+from sparkle.CLI.help.nicknames import resolve_object_name
+from sparkle.instance import resolve_instance_name